feat(api): rename validations to Validations
