+++ apps/api/app/Controllers/Http/Admin/AccountsController.ts
@@ -9,8 +9,8 @@ import {
   genderRules,
   userIdRules,
   userIdUpdateRules,
-} from 'App/validations/account'
-import { passwordRules } from 'App/validations/user'
+} from 'App/Validations/account'
+import { passwordRules } from 'App/Validations/user'
 
 export default class AccountsController {
   /**
